feat: update render methods and folder structure
