--- src/components/row/index.ts
@@ -1,3 +1,0 @@
-import Row from './row';
-
-export default Row;
