+++ src/components/table-cell/table-cell.scss
@@ -1,0 +1,3 @@
+.table-cell {
+  width: fit-content;
+}
