--- src/components/table-cell/teble-cell.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-interface TableCellProps {
-  children: string | number;
-  scope?: string;
-}
-
-const TableCell: React.FC<TableCellProps> = ({ children }) => (
-  <td>{children}</td>
-);
-
-export default TableCell;
