--- src/components/table-head/head.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const Head: React.FC = () => (
-  <thead className="thead-dark">
-    <tr>
-      <th scope="col">id</th>
-      <th scope="col">Name</th>
-      <th scope="col">Job title</th>
-      <th scope="col">Company</th>
-      <th scope="col">Salary (yearly)</th>
-      <th scope="col">Country</th>
-      <th scope="col">Email</th>
-      <th scope="col">Phone number</th>
-    </tr>
-  </thead>
-);
-
-export default Head;
