+++ src/components/table-head/table-head.tsx
@@ -1,0 +1,30 @@
+import React from 'react';
+
+import TableCell from '../table-cell';
+
+const Head: React.FC = () => {
+  const headings = [
+    'id',
+    'Name',
+    'Job title',
+    'Company',
+    'Salary (yearly)',
+    'Country',
+    'Email',
+    'Phone number',
+  ];
+
+  return (
+    <thead className="thead-dark">
+      <tr>
+        {
+          headings.map((heading) => (
+            <TableCell scope="col" th key={heading}>{heading}</TableCell>
+          ))
+        }
+      </tr>
+    </thead>
+  );
+};
+
+export default Head;
