+++ src/components/table-row/index.ts
@@ -1,0 +1,3 @@
+import TableRow from './table-row';
+
+export default TableRow;
